dental chairs part 1
debug:

- Edit chair availbility per shift (for today)
- edit chair availability to Under Maintenance
